requirements.txt
asgiref==3.5.2
Django==4.1.2
sqlparse==0.4.3
django-ckeditor==6.5.1
django-uuslug==2.0.0
Pillow==9.2.0
python-decouple==3.6
django-cleanup==6.0.0